+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The contents of this changelog are focused on changes to overall behaviour of th
 using its classes. Internal changes that don't affect this behaviour are ommitted, or simply noted as 'Minor internal
 changes', 'Minor performance optimizations', etc.
 
+## v1.1.0 - [UNRELEASED]
+### Added
+ - `PclValidations::class` - Class of common validation based functions
+ - `PclValidations::ValidateUrl()` - Validates a string using PHP's `filter_var()` function with the
+    `FILTER_VALIDATE_URL` filter and a regex pattern to validate specific protocols.
+
 ## v1.0.0 - 2017-07-25
 ### Added
  - `PclExceptionWithCode::class` - Extension of `\Exception` to provide predefined 'exception codes' for debugging/error